# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Full Stack FastAPI Project", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/accounts": {
      "post": {
        "tags": ["accounts"],
        "summary": "Create Account",
        "operationId": "accounts-create_account",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AccountCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["accounts"],
        "summary": "Read Accounts",
        "operationId": "accounts-read_accounts",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountsPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{account_id}": {
      "put": {
        "tags": ["accounts"],
        "summary": "Update Account",
        "operationId": "accounts-update_account",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AccountUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios": {
      "post": {
        "tags": ["portfolios"],
        "summary": "Create Portfolio",
        "operationId": "portfolios-create_portfolio",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PortfolioCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PortfolioPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["portfolios"],
        "summary": "Read Portfolios",
        "operationId": "portfolios-read_portfolios",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "uuid" },
                { "type": "null" }
              ],
              "title": "Account Id"
            }
          },
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PortfoliosPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolios/{portfolio_id}": {
      "put": {
        "tags": ["portfolios"],
        "summary": "Update Portfolio",
        "operationId": "portfolios-update_portfolio",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PortfolioUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PortfolioPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "account_type": { "$ref": "#/components/schemas/AccountType" },
          "institution_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Institution Name"
          },
          "account_mask": {
            "anyOf": [
              { "type": "string", "maxLength": 32 },
              { "type": "null" }
            ],
            "title": "Account Mask"
          },
          "base_currency": {
            "type": "string",
            "maxLength": 8,
            "title": "Base Currency",
            "default": "USD"
          },
          "notes": {
            "anyOf": [
              { "type": "string", "maxLength": 1000 },
              { "type": "null" }
            ],
            "title": "Notes"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": ["name", "account_type", "institution_name"],
        "title": "AccountCreate"
      },
      "AccountPublic": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "account_type": { "$ref": "#/components/schemas/AccountType" },
          "institution_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Institution Name"
          },
          "account_mask": {
            "anyOf": [
              { "type": "string", "maxLength": 32 },
              { "type": "null" }
            ],
            "title": "Account Mask"
          },
          "base_currency": {
            "type": "string",
            "maxLength": 8,
            "title": "Base Currency",
            "default": "USD"
          },
          "notes": {
            "anyOf": [
              { "type": "string", "maxLength": 1000 },
              { "type": "null" }
            ],
            "title": "Notes"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "account_type",
          "institution_name",
          "id",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "AccountPublic"
      },
      "AccountType": {
        "type": "string",
        "enum": ["brokerage", "bank"],
        "title": "AccountType"
      },
      "AccountUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              { "type": "string", "maxLength": 255, "minLength": 1 },
              { "type": "null" }
            ],
            "title": "Name"
          },
          "account_type": {
            "anyOf": [
              { "$ref": "#/components/schemas/AccountType" },
              { "type": "null" }
            ]
          },
          "institution_name": {
            "anyOf": [
              { "type": "string", "maxLength": 255, "minLength": 1 },
              { "type": "null" }
            ],
            "title": "Institution Name"
          },
          "account_mask": {
            "anyOf": [
              { "type": "string", "maxLength": 32 },
              { "type": "null" }
            ],
            "title": "Account Mask"
          },
          "base_currency": {
            "anyOf": [{ "type": "string", "maxLength": 8 }, { "type": "null" }],
            "title": "Base Currency"
          },
          "notes": {
            "anyOf": [
              { "type": "string", "maxLength": 1000 },
              { "type": "null" }
            ],
            "title": "Notes"
          },
          "is_active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "AccountUpdate"
      },
      "AccountsPublic": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/AccountPublic" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "AccountsPublic"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortfolioCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "description": {
            "anyOf": [
              { "type": "string", "maxLength": 1000 },
              { "type": "null" }
            ],
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": ["name", "account_id"],
        "title": "PortfolioCreate"
      },
      "PortfolioPublic": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "description": {
            "anyOf": [
              { "type": "string", "maxLength": 1000 },
              { "type": "null" }
            ],
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "account_id",
          "id",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "PortfolioPublic"
      },
      "PortfolioUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              { "type": "string", "maxLength": 255, "minLength": 1 },
              { "type": "null" }
            ],
            "title": "Name"
          },
          "account_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Account Id"
          },
          "description": {
            "anyOf": [
              { "type": "string", "maxLength": 1000 },
              { "type": "null" }
            ],
            "title": "Description"
          },
          "is_active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "PortfolioUpdate"
      },
      "PortfoliosPublic": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/PortfolioPublic" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "PortfoliosPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}